Move classes to otel-extension module
