+++ agent/src/test/scala/io/scalac/mesmer/agent/utils/InstallAgent.scala
@@ -8,14 +8,13 @@ import org.scalatest.BeforeAndAfterAll
 import org.scalatest.TestSuite
 
 import io.scalac.mesmer.agent.Agent
-import io.scalac.mesmer.agent.akka.actor.AkkaActorAgent
 import io.scalac.mesmer.agent.akka.persistence.AkkaPersistenceAgent
 import io.scalac.mesmer.agent.util.i13n.InstrumentModuleFactory
 import io.scalac.mesmer.agent.utils.InstallAgent.allInstrumentations
 import io.scalac.mesmer.core.module.MesmerModule
 
 object InstallAgent {
-  def allInstrumentations: Agent = AkkaActorAgent.agent ++ AkkaPersistenceAgent.agent
+  def allInstrumentations: Agent = AkkaPersistenceAgent.agent
 }
 
 abstract class InstallAgent extends TestSuite with BeforeAndAfterAll {
