+++ otel-extension/src/main/scala/io/scalac/mesmer/otelextension/instrumentations/akka/actor/impl/LocalActorRefProviderAdvice.scala
@@ -1,8 +1,8 @@
-package io.scalac.mesmer.agent.akka.actor.impl
+package io.scalac.mesmer.otelextension.instrumentations.akka.actor.impl
 
 import akka.actor.ActorRef
 import akka.actor.ActorSystem
-import akka.actor.typed.scaladsl.adapter._
+import akka.actor.typed.scaladsl.adapter.ClassicActorSystemOps
 import net.bytebuddy.asm.Advice.Argument
 import net.bytebuddy.asm.Advice.OnMethodExit
 import net.bytebuddy.asm.Advice.Return
